removendo duplicatas
