+++ biblioteca/app/src/main/java/com/hylsonk/biblioteca/Repository/RepositoryBooks.kt
@@ -3,7 +3,12 @@ package com.hylsonk.biblioteca.Repository
 import com.hylsonk.biblioteca.Model.Book
 
 class RepositoryBooks {
-    private val books: List<Book>
+    private val books = arrayOf(
+        "Prisioneiros da Mente",
+        "O Poder do Hábito",
+        "Harry Potter e as Relíquias da Morte",
+        "A Guerra dos Tronos"
+    )
 
     fun getBooks(): Array<String>{
         return books
